allow the user to set the android-sdk location
