+++ packages/flutter_tools/lib/src/commands/config.dart
@@ -20,6 +20,7 @@ class ConfigCommand extends FlutterCommand {
       negatable: false,
       help: 'Clear the saved development certificate choice used to sign apps for iOS device deployment.');
     argParser.addOption('gradle-dir', help: 'The gradle install directory.');
+    argParser.addOption('android-sdk', help: 'The Android SDK directory.');
     argParser.addOption('android-studio-dir', help: 'The Android Studio install directory.');
     argParser.addFlag('machine',
       negatable: false,
@@ -72,6 +73,9 @@ class ConfigCommand extends FlutterCommand {
 
     if (argResults.wasParsed('gradle-dir'))
       _updateConfig('gradle-dir', argResults['gradle-dir']);
+
+    if (argResults.wasParsed('android-sdk'))
+      _updateConfig('android-sdk', argResults['android-sdk']);
 
     if (argResults.wasParsed('android-studio-dir'))
       _updateConfig('android-studio-dir', argResults['android-studio-dir']);
